Conflicts:
	src/presentation/UC1.fxml

+++ src/logic/FTPController.java
@@ -18,5 +18,5 @@ public interface FTPController {
 	
 	public boolean accepterPris();
 	
-	public Profil anmodOmProfil(double kundeNummer) throws SQLException;
+	public Profil anmodOmProfil(long kundeNummer) throws SQLException;
 }
