+++ src/persistence/ProfilKartotekImpl.java
@@ -13,7 +13,7 @@ public class ProfilKartotekImpl implements ProfilKartotek {
 	final static String SELECT_PROFIL = "SELECT * FROM Profiler WHERE kundeNummer = ?";
 
 	@Override
-	public Profil anmodOmProfil(double kundeNummer) throws SQLException {
+	public Profil anmodOmProfil(long kundeNummer) throws SQLException {
 		DataAccessForSQL da = new DataAccessForSQL();
 		Connection connection = da.getConnection();
 		PreparedStatement ps = connection.prepareStatement(SELECT_PROFIL);
@@ -24,7 +24,7 @@ public class ProfilKartotekImpl implements ProfilKartotek {
 		while(resultset.next()){
 			String fuldtNavn = resultset.getString("fuldt_navn");
 			String email = resultset.getString("email");
-			double tlfNummer = resultset.getDouble("tlf_nummer");
+			long tlfNummer = resultset.getLong("tlf_nummer");
 			
 			profil.setKundeNummer(kundeNummer);
 			profil.setFuldtNavn(fuldtNavn);
